pyproject.toml
[project]
name = "typeform-rag-chatbot"
version = "0.1.0"
description = "AI-powered chatbot for Typeform Help Center using RAG"
authors = [
    {name = "Typeform Data Science Team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core web framework
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pydantic==2.5.0",
    # LLM and Embeddings
    "openai>=1.6.1",
    "langchain==0.1.0",
    "langchain-openai==0.0.2",
    # Vector Database
    "pinecone-client==3.0.0",
    # Configuration and utilities
    "python-multipart==0.0.6",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.14.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "data", "scripts"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
